replacement effect skipping first turn draw step
ContinuousEffect implementation is pretty thin. Probably tackle that
with more substance in the layer system.

Really need to streamline the GameEvent/GameAction separation. I think
there's a need for a series of events that are always called after an
event in addition to optional Options, like an "Ok" prompt to pass turn.